HDFS-1473. Refactor image loading and saving from image storage management
