+++ src/java/org/apache/hadoop/hdfs/server/namenode/FSNamesystem.java
@@ -4664,7 +4664,7 @@ public class FSNamesystem implements FSConstants, FSNamesystemMBean, FSClusterSt
                                   " but is not under construction.");
           }
           INodeFileUnderConstruction cons = (INodeFileUnderConstruction) node;
-          FSImage.writeINodeUnderConstruction(out, cons, path);
+          FSImageWriter.writeINodeUnderConstruction(out, cons, path);
         }
       }
     }
